split cpu parts in permute_pooled_embedding_ops for cpu_only
