+++ fbgemm_gpu/CMakeLists.txt
@@ -225,6 +225,7 @@ set(fbgemm_gpu_sources_cpu
     src/cpu_utils.cpp
     src/input_combine_cpu.cpp
     src/layout_transform_ops_cpu.cpp
+    src/permute_pooled_embedding_ops_cpu.cpp
     src/quantize_ops_cpu.cpp
     src/sparse_ops_cpu.cpp)
