+++ fbgemm_gpu/fbgemm_gpu/permute_pooled_embedding_modules.py
@@ -17,6 +17,9 @@ try:
     from fbgemm_gpu import open_source  # noqa: F401
 except Exception:
     torch.ops.load_library(
+        "//deeplearning/fbgemm/fbgemm_gpu:permute_pooled_embedding_ops_cpu"
+    )
+    torch.ops.load_library(
         "//deeplearning/fbgemm/fbgemm_gpu:permute_pooled_embedding_ops_gpu"
     )
