--- fbgemm_gpu/include/fbgemm_gpu/permute_pooled_embs_function.h
@@ -1,6 +1,0 @@
-/*
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- * All rights reserved.
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree.
- */
